Rewrite import map urls for custom domains. Resolves #59
