+++ .changeset/lazy-peas-argue.md
@@ -1,0 +1,5 @@
+---
+"baseplate-cloudflare-worker": minor
+---
+
+Rewrite urls in import maps to use custom domains
